Clean draggable options
I was actually trying to find a way to set boundaries to the drag area, so that you can't drag the items around the whole screen but seems like vue draggable and the underlying sortable.js doesn't support this. It's not too bad but in buckets it looks like you could drag and drop from one to the other. Also noticed dragging up in forms (for example in a slideshow compontent) is not scrolling the page up when dragging to the top, which seems to be caused by the sticky navigation.