Drop laravel-mix requirement on host projects
To avoid conflicts with arbitrary npm setups in projects using this package (like a project running on Webpack 4, which is not compatible with Laravel Mix as of April 2018), I removed the requirement to install it and modfied the npm scripts to use a simple copy command.

This means the cms build is fully independant from the project build. It also gives up a slight performance boost in HMR mode when developing.

Next step would be to turn those scripts into a .js file, that we could run from an artisan command shipped with the toolkit. This would completely remove the friction to get started working for a developer new to this package.