Add linguist override in .gitattribute
GitHub "Language Savant" tool Linguist is reporting Twill as mainly a Vue project because of its statistics. While Twill contains a lot of Vue SFCs, it is a Laravel package, not a Vue library. In order to have PHP showing as the main language, we ignore Vue components using Linguist's overrides.