Keep file library uuid backwards compatible by introducing a new twill config key
twill.file_library.prefix_uuid_with_local_path defaults to false. When true, the uuid generated and saved in database will behave like the media library when used with the local driver, which means it is prefixed by the folder name of the file library local path. This is meant to keep the next release out of breaking changes before merging #173 and more cleanup in 1.3.