Change the way of maintaining Twill version number (#404)
* Removed the VERSION file, use the constant VERSION in TwillServiceProvider to store it instead

* Remove version method type-hinting

We do not use return type-hinting anywhere else in Twill's codebase but have started to add doc blocks helping IDEs with return types recently. Since this method already has a doc block about the return type, the hint is not necessary. There's definitely a debate on both approaches and while type-hinting might be an approach we implement globally in the future, this is not something we do at the moment.