Fix #605
With the simplification of the exception handler in #561 the override of the invalidJson method was dropped by mistake. It is currently necessary as the frontend expects errors only in a validation exception response. In future improvments it would be better to keep the default Laravel response format and update the frontend to read one level deeper.