Fix tests
See 1bf93ada. Login has its own test so we do not need to check the status of the login request on every test. It appears that on a first run the Testbench created Laravel app does not pick up the admin.php file resulting in a route missing when being redirected to the dashboard after login.