adds `direction` option to form inputs
It maybe that you content entry is in a RTL
language, such as Arabic, but, you want to LTR
content such as a URL or an email address.

Or, it maybe your content entry is in a LTR
language and you need a single RTL entry field.

This commit adds `direction` for this purpose:

```
Input::make()
    ->name('url')
    ->label(twillTrans('Url'))
    ->direction('ltr')
```